requirements.txt
python=3.9.7
setuptools=58.0.4
yaml=0.2.5
h5py==3.7.0
matplotlib==3.5.2
nltk==3.7
numpy==1.22.3
pandas==1.4.2
pillow==9.2.0
scikit-learn==1.1.2
scipy==1.8.1
seaborn==0.11.2
setproctitle==1.2.3
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
wandb==0.13.2
yacs==0.1.8